CONSOLE-5021: Additional Cypress config dedupe
