+++ frontend/packages/knative-plugin/integration-tests/cypress.config.js
@@ -1,45 +1,14 @@
-const { defineConfig } = require('cypress');
-const commonConfig = require('@console/cypress-integration-tests/cypress-common-config');
+const { defineConfig } = require('@console/cypress-integration-tests/cypress-common-config');
 
 module.exports = defineConfig({
-  ...commonConfig,
-  defaultCommandTimeout: 40000,
-  animationDistanceThreshold: 40,
-  execTimeout: 270000,
-  pageLoadTimeout: 100000,
-  requestTimeout: 15000,
-  responseTimeout: 15000,
   fixturesFolder: 'testData',
-  video: true,
-  reporter: '../../../node_modules/cypress-multi-reporters',
-  reporterOptions: {
-    configFile: 'reporter-config.json',
-  },
-  screenshotsFolder: '../../../gui_test_screenshots/cypress/screenshots',
-  videosFolder: '../../../gui_test_screenshots/cypress/videos',
   env: {
     TAGS:
       '(@knative or @knative-admin or @knative-camelk or @knative-eventing or @knative-kafka or @knative-serverless) and (@pre-condition or @smoke or @regression) and not (@manual or @to-do or @un-verified or @broken-test)',
     NAMESPACE: 'aut-knative',
   },
-  retries: {
-    runMode: 1,
-    openMode: 0,
-  },
   e2e: {
-    // We've imported your old cypress plugins here.
-    // You may want to clean this up later by importing these.
-    setupNodeEvents(on, config) {
-      // eslint-disable-next-line global-require
-      return require('./../../dev-console/integration-tests/plugins/index.js')(on, config);
-    },
     specPattern: 'features/**/*.{feature,features}',
     supportFile: 'support/commands/index.ts',
-    baseUrl: 'http://localhost:9000',
-    testIsolation: false,
-    experimentalMemoryManagement: true,
-    numTestsKeptInMemory: 5,
-    injectDocumentDomain: true,
-    userAgent: 'ConsoleIntegrationTestEnvironment',
   },
 });
