+++ frontend/packages/shipwright-plugin/integration-tests/cypress.config.js
@@ -1,44 +1,13 @@
-const { defineConfig } = require('cypress');
-const commonConfig = require('@console/cypress-integration-tests/cypress-common-config');
+const { defineConfig } = require('@console/cypress-integration-tests/cypress-common-config');
 
 module.exports = defineConfig({
-  ...commonConfig,
-  defaultCommandTimeout: 40000,
-  animationDistanceThreshold: 20,
-  execTimeout: 90000,
-  pageLoadTimeout: 90000,
-  requestTimeout: 15000,
-  responseTimeout: 15000,
   fixturesFolder: 'testData',
-  video: true,
-  reporter: '../../../node_modules/cypress-multi-reporters',
-  reporterOptions: {
-    configFile: 'reporter-config.json',
-  },
-  screenshotsFolder: '../../../gui_test_screenshots/cypress/screenshots',
-  videosFolder: '../../../gui_test_screenshots/cypress/videos',
   env: {
     TAGS: '(@smoke or @regression) and not (@manual or @to-do or @un-verified or @broken-test)',
     NAMESPACE: 'aut-shipwright',
   },
-  retries: {
-    runMode: 1,
-    openMode: 0,
-  },
   e2e: {
-    // We've imported your old cypress plugins here.
-    // You may want to clean this up later by importing these.
-    setupNodeEvents(on, config) {
-      // eslint-disable-next-line global-require
-      return require('./../../dev-console/integration-tests/plugins/index.js')(on, config);
-    },
     specPattern: 'features/**/*.{feature,features}',
     supportFile: 'support/commands/index.ts',
-    baseUrl: 'http://localhost:9000',
-    testIsolation: false,
-    experimentalMemoryManagement: true,
-    numTestsKeptInMemory: 5,
-    injectDocumentDomain: true,
-    userAgent: 'ConsoleIntegrationTestEnvironment',
   },
 });
